requirements.txt
git==3.1.32
pygithub==2.6.1
langchain==0.1.5
langchain-community==0.0.19
langchain-pinecone==0.2.3
pinecone-client==6.0.0
tiktoken==0.5.2
openai==1.13.3
python-dotenv==1.0.1
sentence-transformers==2.3.1
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-java==0.20.1
tree-sitter-cpp==0.20.0
tree-sitter-javascript==0.20.1
tree-sitter-go==0.20.0
tree-sitter-html==0.20.0 